Switch to unmanaged containers in Zig script (refactoring)
PR #750 (bdwgc).

This is in preparation for the upcoming Zig 0.15 release.
Managed containers (storing an allocator) will go away, unmanaged ones
(passing an allocator to functions that need to allocate) are becoming
the default.  What is called `std.ArrayListUnmanaged` up until Zig 0.14
has been renamed to `std.ArrayList` in Zig 0.15 with the old name still
available as a deprecated alias.

This change is backwards compatible as both types of containers have
existed for a long time.

* build.zig (build, addTest, addTestExt): Use
`std.ArrayListUnmanaged()` instead of `std.ArrayList()`.
* build.zig (build): Use `b.allocator`.